Move formatBytes into CoreUtils.java
