+++ java/clients/src/test/java/sleeper/clients/util/ClientUtilsTest.java
@@ -26,7 +26,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.clients.util.ClientUtils.formatBytes;
 import static sleeper.clients.util.Command.command;
 import static sleeper.clients.util.CommandPipeline.pipeline;
 
@@ -180,64 +179,6 @@ class ClientUtilsTest {
             // When/Then
             assertThat(ClientUtils.runCommandInheritIO(command))
                     .isEqualTo(new CommandPipelineResult(1, 0));
-        }
-    }
-
-    @Nested
-    @DisplayName("Format bytes as string")
-    class FormatBytesAsString {
-        @Test
-        void shouldFormatNumberOfBytesBelow1KB() {
-            assertThat(formatBytes(123L))
-                    .isEqualTo("123B");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAsKB() {
-            assertThat(formatBytes(1_234L))
-                    .isEqualTo("1.2KB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAsKBWithRounding() {
-            assertThat(formatBytes(5_678L))
-                    .isEqualTo("5.7KB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesEqualTo1KB() {
-            assertThat(formatBytes(1_000L))
-                    .isEqualTo("1.0KB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAs10KB() {
-            assertThat(formatBytes(10_000L))
-                    .isEqualTo("10.0KB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAsMB() {
-            assertThat(formatBytes(1_234_000L))
-                    .isEqualTo("1.2MB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAsGB() {
-            assertThat(formatBytes(1_234_000_000L))
-                    .isEqualTo("1.2GB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAbove1TB() {
-            assertThat(formatBytes(1_234_000_000_000L))
-                    .isEqualTo("1TB");
-        }
-
-        @Test
-        void shouldFormatNumberOfBytesAbove1000TB() {
-            assertThat(formatBytes(1_234_000_000_000_000L))
-                    .isEqualTo("1,234TB");
         }
     }
 }
